Move dictCp command to separate file
